tartarus-bot-example/pyproject.toml
[project]
name = "tartarus-bot-example"
version = "0.1.0"
description = ""
authors = [
    {name = "SubspacedBoy",email = "ssb@subjugated.club"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "protobuf (>=6.30.1,<7.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "stomp-py (>=8.2.0,<9.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "stomper (>=0.4.3,<0.5.0)",
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "flatbuffers (>=25.2.10,<26.0.0)"
]

[tool.poetry]
packages = [{include = "tartarus_bot_example", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
